Next js 14 and app router setup
