--- next.config.js
@@ -1,9 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-  images: {
-    domains: ["admin.blazecodes.com", 'blazecodes.com', "system.pradhansaroj.com.np", 'wwww.blazecodes.com', "127.0.0.1", "images.unsplash.com"],
-  },
-};
-
-module.exports = nextConfig;
