+++ next.config.mjs
@@ -1,0 +1,8 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+    images: {
+        domains: ["admin.blazecodes.com", 'blazecodes.com', "system.pradhansaroj.com.np", 'wwww.blazecodes.com', "127.0.0.1", "images.unsplash.com"],
+      },
+};
+
+export default nextConfig;
