+++ src/app/blogs/page.js
@@ -1,15 +1,19 @@
+"use client";
+
 import Image from "next/image";
-import Link from "next/link";
 import React from "react";
-import SrzLayout from "../components/SrzLayout";
-import data from "../src/data";
-import RevealOnScroll from "../components/Reveal";
-export default function blogs({ posts }) {
-  // console.log(posts);
+import SrzLayout from "@/components/SrzLayout";
+import RevealOnScroll from "@/components/Reveal";
+import useSWR from "swr";
+
+export default function Page() {
+    const fetcher = (...args) => fetch(...args).then(res => res.json())
+    const { data:posts , isLoading, error } = useSWR("https://system.pradhansaroj.com.np/api/srz-portfolio-post-list", fetcher);
+    // console.log(posts, isLoading, error)
   return (
     <SrzLayout title="Blogs">
       <div className="grid xl:grid-cols-4 lg:grid-cols-3 md:grid-cols-2 md:gap-5">
-        {posts.length >= 1 ? (
+        {posts && posts.length >= 1 ? (
           posts.map((post, index) => (
             <RevealOnScroll
               key={index}
@@ -63,25 +67,4 @@ export default function blogs({ posts }) {
   );
 }
 
-export async function getStaticProps() {
-  let posts = [];
-  try {
-    var res = await fetch(`https://system.pradhansaroj.com.np/api/srz-portfolio-post-list`);
-    posts = await res.json();
-  } catch (e) {
-    posts = [
-      {
-        title: "Fetch Error",
-        thumbnail: "/fail",
-        category: { title: "fetch error" },
-      },
-    ];
-  }
 
-  return {
-    props: {
-      posts,
-    },
-    revalidate: 10, // Revalidate every 10 seconds with new data.
-  };
-}
