+++ src/app/style.css
@@ -1,0 +1,363 @@
+:root {
+    --primary: #6244c5;
+    --secondary: #f5a14d;
+    --light: #fafafb;
+    --dark: #12141d;
+  }
+  
+  body {
+    background-color: var(--dark) !important;
+    font-family: "Poppins", sans-serif !important;
+  }
+  a {
+    text-decoration: none;
+  }
+  p {
+    color: gray;
+  }
+  img {
+    border-radius: 8px;
+  }
+  /* custom scrollbar  */
+  ::-webkit-scrollbar {
+    width: 8px;
+    height: 8px;
+  }
+  
+  /* Track */
+  ::-webkit-scrollbar-track {
+    background: rgb(17, 24, 39);
+  }
+  
+  /* Handle */
+  ::-webkit-scrollbar-thumb {
+    background: #607197;
+    border-radius: 10px;
+  }
+  
+  /* Handle on hover */
+  ::-webkit-scrollbar-thumb:hover {
+    background: #314163;
+  }
+  
+  .object-cover {
+    object-fit: cover;
+  }
+  
+  .text-justify {
+    text-align: justify;
+  }
+  .pointer {
+    cursor: pointer;
+  }
+  /* Set the initial opacity to 0 and transition to 0.5 seconds */
+  .reveal-on-scroll {
+    opacity: 0;
+    transform: translateY(50px);
+    transition: all 0.5s ease-in-out;
+  }
+  
+  /* When the element becomes visible, change the opacity to 1 */
+  .reveal-on-scroll.visible {
+    opacity: 1;
+    transform: translateY(0);
+  }
+  
+  .reveal-on-scroll.hidden {
+    opacity: 0;
+    transition: all 0.5s ease-in-out;
+  }
+  .debug {
+    border: 2px solid red;
+  }
+  .back-to-top {
+    position: fixed;
+    display: none;
+    right: 40px;
+    bottom: 80px;
+    z-index: 99;
+  }
+  .my-6 {
+    margin-top: 6rem;
+    margin-bottom: 6rem;
+  }
+  
+  .py-6 {
+    padding-top: 6rem;
+    padding-bottom: 6rem;
+  }
+  
+  .container {
+    width: 100%;
+    margin-right: auto;
+    margin-left: auto;
+    padding-right: 16px;
+    padding-left: 16px;
+  }
+  
+  @media (min-width: 540px) {
+    .container {
+      max-width: 540px;
+    }
+  }
+  
+  @media (min-width: 720px) {
+    .container {
+      max-width: 720px;
+    }
+  }
+  
+  @media (min-width: 960px) {
+    .container {
+      max-width: 960px;
+    }
+  }
+  
+  @media (min-width: 1140px) {
+    .container {
+      max-width: 1140px;
+    }
+  }
+  
+  @media (min-width: 1320px) {
+    .container {
+      max-width: 1320px;
+    }
+  }
+  
+  /*** Spinner ***/
+  #spinner {
+    opacity: 0;
+    visibility: hidden;
+    transition: opacity 0.5s ease-out, visibility 0s linear 0.5s;
+    z-index: 99999;
+  }
+  
+  #spinner.show {
+    transition: opacity 0.5s ease-out, visibility 0s linear 0s;
+    visibility: visible;
+    opacity: 1;
+  }
+  /* custom bg color  */
+  .bg-primary {
+    background-color: #6244c5 !important;
+  }
+  .bg-secondary {
+    background-color: #bd8204 !important;
+  }
+  .bg-success {
+    background-color: #198754 !important;
+  }
+  .bg-info {
+    background-color: #0dcaf0 !important;
+  }
+  .bg-warning {
+    background-color: #ffc107 !important;
+  }
+  .bg-danger {
+    background-color: #dc3545 !important;
+  }
+  .bg-light {
+    background-color: #fafafb !important;
+  }
+  .bg-dark {
+    background-color: #12141d !important;
+  }
+  
+  /*** Button ***/
+  .btn {
+    font-weight: 500;
+    transition: 0.5s;
+  }
+  
+  a {
+    color: var(--primary);
+  }
+  .text-primary {
+    color: var(--primary) !important;
+  }
+  
+  /*** Navbar ***/
+  
+  .nav-item {
+    background: linear-gradient(var(--primary) 0 0) bottom / var(--d, 0) 3px
+      no-repeat;
+    transition: 0.5s;
+  }
+  .nav-item:hover {
+    --d: 100%;
+  }
+  .active-nav-item {
+    background: linear-gradient(var(--secondary) 0 0) bottom / 100% 3px no-repeat;
+  }
+  .mobile-nav-item {
+    transition: 0.5s ease-in-out;
+  }
+  
+  .active-mobile-nav-item {
+    background-color: var(--primary);
+    filter: brightness(0.85);
+    transition: 0.5s ease-in-out;
+  }
+  /* .navbar.fixed-top {
+      display: none;
+      transition: .5s;
+  } */
+  
+  /* .navbar .navbar-nav .nav-link {
+      padding: 20px 15px;
+      color: var(--dark);
+      font-size: 18px;
+      font-weight: 600;
+      outline: none;
+  }
+  
+  .navbar .navbar-nav .nav-link:hover,
+  .navbar .navbar-nav .nav-link.active {
+      color: var(--primary);
+  } */
+  
+  /* @media (max-width: 991.98px) {
+      .navbar .navbar-nav .nav-link  {
+          padding: 10px 0;
+      }
+  } */
+  .logo-title {
+    font-family: "Righteous", cursive !important;
+    font-size: 40px;
+  }
+  
+  /*** Header ***/
+  #home {
+    min-height: 100vh;
+    background-position: center;
+    background-size: cover;
+    display: flex;
+    align-items: center;
+  }
+  .home-cover {
+    position: absolute;
+    top: 3px;
+    bottom: 0px;
+    left: 0px;
+    right: 0px;
+    z-index: 1;
+  }
+  
+  .home-container {
+    position: absolute;
+    z-index: 1;
+    width: 80vw;
+    left: 10vw;
+    top: 50%;
+    transform: translateY(-50%);
+  }
+  .typed-cursor {
+    font-size: 30px;
+    color: var(--dark);
+  }
+  
+  .content-wrapper {
+    padding: 100px 0;
+    min-height: 90vh;
+  }
+  
+  /*** Service ***/
+  .service-card {
+    /* display: flex; */
+  }
+  .service-img-holder {
+    position: relative;
+    height: 200px;
+    width:200px;
+    object-fit: cover;
+  }
+  .service-card p {
+    font-size: 0.9rem;
+    text-align: justify;
+  }
+  .service-card h5 {
+    font-weight: 400;
+  }
+  /*** work Portfolio ***/
+  #portfolio-filters li {
+    display: inline-block;
+    font-weight: 500;
+    cursor: pointer;
+    transition: 0.5s;
+    border-bottom: 2px solid transparent;
+  }
+  
+  #portfolio-filters li:hover,
+  #portfolio-filters li.active {
+    color: var(--secondary);
+    padding-bottom: 5px;
+    position: relative;
+  }
+  
+  #portfolio-filters li.active:after {
+    content: "";
+    background: rgb(252, 249, 249);
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    height: 2px;
+    width: 50%;
+  }
+  
+  .portfolio-img {
+    position: relative;
+    height: 250px;
+    width: auto;
+  }
+  
+  .portfolio-btn {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    background: rgba(98, 68, 197, 0.9);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 30px;
+    padding-bottom: 100px !important;
+    opacity: 0;
+    transition: 0.5s;
+  }
+  
+  .portfolio-item:hover .portfolio-btn {
+    opacity: 1;
+    padding-bottom: 30px !important;
+  }
+  
+  .portfolio-hover-title {
+    color: white;
+    font-size: 30px;
+    text-align: center;
+    font-weight: bold;
+  }
+  
+  .portfolio-hover-view {
+    position: absolute;
+    top: 60%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    color: white;
+  }
+  
+  .job-detail__image-holder {
+    height: 350px;
+  }
+  
+  @media screen and (max-width: 768px) {
+    .job-detail__image-holder {
+      height: 200px;
+    }
+  
+    .content-wrapper {
+      padding: 80px 0;
+    }
+  }
+
