+++ src/data/workData.json
@@ -1,5 +1,4 @@
-{
-  "works": [
+[
     {
       "title": "Kumari Job",
       "slug": "kumari-job-web",
@@ -133,4 +132,3 @@
       "is_featured": false
     }
   ]
-}
